♻️ DGSB7B code: split maximum assimilation artifact gates
